fixing imports from refactor of api spec
